Equipamentos e Disciplina endpoints.
* Equipamentos
** Adicionado validação para as requisições.

* Disciplina
** Adicionado controller, service, model, DTO e mapper.

* Mudanças gerais:
** Adicionado validação para os DTOs.
** Mudança dos DTO, de classe para records.
** Mudança dos Integer para Long.

+++ Backend/src/main/java/com/br/sga/domain/Aluno.java
@@ -22,7 +22,7 @@ public class Aluno implements Serializable {
 
     @Id
     @Column(name = "matricula", nullable = false)
-    private Integer matricula;
+    private Long matricula;
 
     @Column(name = "nome")
     private String nome;
