+++ Backend/src/main/java/com/br/sga/domain/Coordenadoria.java
@@ -1,13 +1,7 @@
 package com.br.sga.domain;
 
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -27,7 +21,7 @@ public class Coordenadoria implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_coordenadoria")
     @SequenceGenerator(name = "seq_coordenadoria", sequenceName = "seq_coordenadoria", allocationSize = 1)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "nome")
     private String nome;
