+++ Backend/src/main/java/com/br/sga/domain/Local.java
@@ -1,13 +1,7 @@
 package com.br.sga.domain;
 
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -27,11 +21,11 @@ public class Local implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_local")
     @SequenceGenerator(name = "seq_local", sequenceName = "seq_local", allocationSize = 1)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "nome", nullable = false)
     private String nome;
 
     @Column(name = "quantidade", nullable = false)
-    private Integer quantidade;
+    private Long quantidade;
 }
