+++ Backend/src/main/java/com/br/sga/repository/AuditoriaAlocacaoRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.AuditoriaAlocacao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditoriaAlocacaoRepository extends JpaRepository<AuditoriaAlocacao, Integer> {
+public interface AuditoriaAlocacaoRepository extends JpaRepository<AuditoriaAlocacao, Long> {
 }
