+++ Backend/src/main/java/com/br/sga/repository/AulaRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.Aula;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AulaRepository extends JpaRepository<Aula, Integer> {
+public interface AulaRepository extends JpaRepository<Aula, Long> {
 }
